fix: r1-fix — split adapter.rs into adapter.rs + ux_sink.rs (≤700 LOC)
adapter.rs 747 → 580 LOC (Channel impl + 10 tests)
ux_sink.rs 186 LOC (UxEventSink impl + 5 tests)

All sub-files now ≤580 LOC.
